% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballon_plot_data_to_result_table.R
\name{ballon_plot_data_to_result_table}
\alias{ballon_plot_data_to_result_table}
\title{Transforms data frame of balloon plot into data frame with weights for each data types in each column}
\usage{
ballon_plot_data_to_result_table(plot_data)
}
\arguments{
\item{plot_data}{output file of make_balloon_plot_data_from_comparison_result() function}
}
\value{
returns a data frame with weights for each data types in each column
}
\description{
Transforms data frame of balloon plot into data frame with weights for each data types in each column
}
\author{
Hakobyan Siras (sirashakobyan@gmail.com)
}
\concept{tumorcomparer}
