% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_freq_alt.R
\name{compute_freq_alt}
\alias{compute_freq_alt}
\title{Computes frequency of alteration (fraction of non-zeroes)}
\usage{
compute_freq_alt(alt_mat)
}
\arguments{
\item{alt_mat}{a matrix or vector of discrete values - 0s represent unaltered, non-zeroes altered}
}
\value{
the fraction of non-zeroes, (i.e. altered features/genes)
}
\description{
Computes frequency of alteration (fraction of non-zeroes)
}
\examples{
set.seed(1)
tmp <- sample(c(0, 1), 100, replace=TRUE)
mat <- matrix(tmp, 10, 10)
compute_freq_alt(mat)

}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
