% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/convert_to_0_to_1_using_xminusmin_by_maxminusmin.R
\name{convert_to_0_to_1_using_xminusmin_by_maxminusmin}
\alias{convert_to_0_to_1_using_xminusmin_by_maxminusmin}
\title{Rescale between 0-1}
\usage{
convert_to_0_to_1_using_xminusmin_by_maxminusmin(x)
}
\arguments{
\item{x}{a numeric vector of values}
}
\value{
a rescaled numeric vector of values
}
\description{
Map the distance matrices for all data types to 0-1
}
