% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyj_graph_maker_from_dist_mat.R
\name{cyj_graph_maker_from_dist_mat}
\alias{cyj_graph_maker_from_dist_mat}
\title{Creates cytoscape.js JSON graph from distance matrix for defined threshold}
\usage{
cyj_graph_maker_from_dist_mat(dist_mat, min_weight)
}
\arguments{
\item{dist_mat}{distance matrix calculated by run_comparison_config_list() function}

\item{min_weight}{numeric value between 0 and 1}
}
\value{
returns cytoscape.js JSON graph
}
\description{
Creates cytoscape.js JSON graph from distance matrix for defined threshold
}
\author{
Hakobyan Siras (sirashakobyan@gmail.com)
}
\concept{tumorcomparer}
