% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_composite_mat_and_gene_weights.R
\name{generate_composite_mat_and_gene_weights}
\alias{generate_composite_mat_and_gene_weights}
\title{Run a comparison between between two cohorts (e.g. cell lines and tumors)}
\usage{
generate_composite_mat_and_gene_weights(
  default_weight,
  known_cancer_gene_weight,
  tumor_file,
  cell_line_file,
  known_cancer_gene_weights_file,
  cancer_specific_gene_weights_file,
  is_discrete,
  distance_similarity_measure
)
}
\arguments{
\item{default_weight}{see run_comparison}

\item{known_cancer_gene_weight}{see run_comparison}

\item{tumor_file}{see run_comparison}

\item{cell_line_file}{see run_comparison}

\item{known_cancer_gene_weights_file}{see run_comparison}

\item{cancer_specific_gene_weights_file}{see run_comparison}

\item{is_discrete}{whether the data is discrete; set automatically by run_comparison for mut, cna, exp}

\item{distance_similarity_measure}{see run_comparison}
}
\value{
a list with multiple items. NOTE: The values of the dist and isomdsfit will
 depend on parameter "distance_similarity_measure".
\itemize{
  \item{"dist_mat"}{a matrix of pairwise distances}
  \item{"isomdsfit"}{a two-column (2-dimension) fitting of the distances reduced to 
  two dimensions via MDS - multidimensional scaling, using the isoMDS function}
  \item{"cor_unweighted"}{a matrix of unweighted pairwise correlations}
  \item{"composite_mat"}{the composite matrix (see Details)}
  \item{"cell_lines_ids"}{a vector of cell line IDs/names}
  \item{"tumors_ids"}{a vector of tumor IDs}
}
}
\description{
Run a comparison between between two cohorts (e.g. cell lines and tumors)
}
\details{
The composite matrix is a single matrix where the columns are samples 
  (i.e. tumors AND cell line IDs) and the rows are an rbind() of mutations 
  (with 1 or 0 outputs for each sample), copy number alterations from 
  GISTIC (with values -2, -1, 0, 1, 2), or gene expression values. Available similarity/distance measures include: 
  \itemize{
    \item{"weighted_correlation"}{Weighted correlation, based on weighted means and standard deviations}
    \item{"generalized_jaccard"}{A weighted distance based on the Jaccard coefficient}
 }
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
