% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_comparison.R
\name{geneset_comparison}
\alias{geneset_comparison}
\title{FIXME}
\usage{
geneset_comparison(
  cancer_type,
  gene_list,
  remove_errored_dataset_comparisons = FALSE
)
}
\arguments{
\item{cancer_type}{FIXME}

\item{gene_list}{FIXME}
}
\value{
FIXME
}
\description{
FIXME
}
\section{TCGA ID and tumor types from TCGA PanCancer annotations}{
TCGA_id_and_tumor_type <- fread("../TC_Data_PanCancer_March2021/TCGA_ID_and_Cancer_Type.txt",check.names=F, nThread = 4)
}

\section{CCLP ID and TCGA tumor type from Cell Model Passport annotations}{
CCLP_TCGA_Types_Combined  <- fread("../TC_Data_PanCancer_March2021/CCLP_ID_and_Cancer_Type.txt",check.names=F, nThread = 4)
\subsection{reading mutation data}{

mut_mat_TCGA_after_Annovar <- fread("../TC_Data_PanCancer_March2021/TCGA_mutation_matrix.txt",check.names=F, nThread = 4)
mut_mat_CCLP_after_Annovar <- fread("../TC_Data_PanCancer_March2021/CCLP_mutation_matrix.txt",check.names=F, nThread = 4)
}

\subsection{unifiing cell line and TCGA genes}{

mut_mat_TCGA_after_Annovar <- mut_mat_TCGA_after_Annovar\link{Gene \%in\% intersect(mut_mat_TCGA_after_Annovar$Gene, mut_mat_CCLP_after_Annovar$Gene)}
mut_mat_CCLP_after_Annovar <- mut_mat_CCLP_after_Annovar\link{Gene \%in\% intersect(mut_mat_TCGA_after_Annovar$Gene, mut_mat_CCLP_after_Annovar$Gene)}
}

\subsection{reading CNV data}{

TCGA_GISTIC_all_data_by_genes <- fread("../TC_Data_PanCancer_March2021/TCGA_GISTIC_all_data_by_genes.txt",check.names=F, nThread = 4)
colnames(TCGA_GISTIC_all_data_by_genes)\link{1} <- "Gene"
CCLP_GISTIC_all_data_by_genes <- fread("../TC_Data_PanCancer_March2021/CCLP_GISTIC_all_data_by_genes.txt",check.names=F, nThread = 4)
colnames(CCLP_GISTIC_all_data_by_genes)\link{1} <- "Gene"
}

\subsection{unifiing cell line and TCGA genes}{

TCGA_GISTIC_all_data_by_genes <- TCGA_GISTIC_all_data_by_genes\link{Gene \%in\% intersect(TCGA_GISTIC_all_data_by_genes$Gene, CCLP_GISTIC_all_data_by_genes$Gene)}
CCLP_GISTIC_all_data_by_genes <- CCLP_GISTIC_all_data_by_genes\link{Gene \%in\% intersect(TCGA_GISTIC_all_data_by_genes$Gene, CCLP_GISTIC_all_data_by_genes$Gene)}
}

\subsection{reading expression data}{

TCGA_Expression_Quantile_Normalized <- fread("../TC_Data_PanCancer_March2021/TCGA_Expression_Quantile_Normalized.txt",check.names=F, nThread = 4)
colnames(TCGA_Expression_Quantile_Normalized)\link{1} <- "Gene"
CCLP_Expression_Quantile_Normalized <- fread("../TC_Data_PanCancer_March2021/CCLP_Expression_Quantile_Normalized.txt",check.names=F, nThread = 4)
}

\subsection{unification is not needed as the data is alredy unified}{
}

\subsection{filtering TCGA samples for all availability of three dat types}{

tcga_mutual_samples <- c("Gene", Reduce(intersect, list(colnames(mut_mat_TCGA_after_Annovar)\link{-1}, colnames(TCGA_GISTIC_all_data_by_genes)\link{-1}, colnames(TCGA_Expression_Quantile_Normalized)\link{-1})))

mut_mat_TCGA_after_Annovar <- mut_mat_TCGA_after_Annovar\link{, ..tcga_mutual_samples}

TCGA_GISTIC_all_data_by_genes <- TCGA_GISTIC_all_data_by_genes\link{, ..tcga_mutual_samples}

TCGA_Expression_Quantile_Normalized <- TCGA_Expression_Quantile_Normalized\link{, ..tcga_mutual_samples}
}

\subsection{filtering CCLP samples for all availability of three dat types}{

cclp_mutual_cell_lines <- c("Gene", Reduce(intersect, list(colnames(mut_mat_CCLP_after_Annovar)\link{-1}, colnames(CCLP_GISTIC_all_data_by_genes)\link{-1}, colnames(CCLP_Expression_Quantile_Normalized)\link{-1})))

mut_mat_CCLP_after_Annovar <- mut_mat_CCLP_after_Annovar\link{, ..cclp_mutual_cell_lines}

CCLP_GISTIC_all_data_by_genes <- CCLP_GISTIC_all_data_by_genes\link{, ..cclp_mutual_cell_lines}

CCLP_Expression_Quantile_Normalized <- CCLP_Expression_Quantile_Normalized\link{, ..cclp_mutual_cell_lines}
}
}

