% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_only_high_level_cnas.R
\name{keep_only_high_level_cnas}
\alias{keep_only_high_level_cnas}
\title{Keep only high-level CNAs (set low-level CNAs to 0)}
\usage{
keep_only_high_level_cnas(cna_mat)
}
\arguments{
\item{cna_mat}{a matrix or vector of 5-valued CNA (copy number alteration) data; 
possible values: -2: Deep deletion, -1: Shallow deletion, 0: Diploid (or default copy number), 
1: Low-level gain, 2: High-level amplification}
}
\value{
a matrix where low-level gains have been set to 0
}
\description{
Keep only high-level CNAs (set low-level CNAs to 0)
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{
tumorcomparer
}

