% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_balloon_plot_data.R
\name{make_balloon_plot_data_from_comparison_result}
\alias{make_balloon_plot_data_from_comparison_result}
\title{Make Balloon Plot from MTC Data structure}
\usage{
make_balloon_plot_data_from_comparison_result(comparison_result)
}
\arguments{
\item{comparison_result}{a data structure generated by run_comparison() function}
}
\value{
a ggplot object
}
\description{
Code to convert the output of TC for a single cancer type to balloon plots. We
  don't use ranks here, since some cell lines will get low ranks even if none have particularly
  low scores (similarly some will get high ranks even if all scores are relatively low). Instead,
  we convert the distance matrices for all tumor-cell line comparisons to 0-1 using min-max
  scaling, and then compute the mean similarity to tumor for each cell line. The overall score is
  just a simiple mean of the data type specific scores right now - could use different weights
}
