% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_balloon_plot_data.R
\name{make_balloon_plot_data_from_mtc}
\alias{make_balloon_plot_data_from_mtc}
\title{Make Balloon Plot from MTC Data structure}
\usage{
make_balloon_plot_data_from_mtc(mtc, cancer_type)
}
\arguments{
\item{mtc}{TODO FIXME DESCRIBE ME EACH COLUMN AND DATA TYPE}

\item{cancer_type}{a cancer type abbreviation found in the MTC column Cell_Line_Cancer_Type}
}
\value{
a ggplot object
}
\description{
Code to convert the processed output of TC over a pan-cancer dataset balloon plots
  for a single cancer type ("current_cancer_type") MTC - The subset of TC's output on a
  pan-cancer dataest which contains results for matched cancer type, i.e. for each cell line, it
  rreports the results when compared to tumors of the same cancer type So, for example, if we run
  TC on 10 cancer types on a pan-cancer type, we get 10 sets of results, with each set containing
  the results of comparing one of the 10 cancer types to all cell lines From this, MTC extracts
  only the results where the cell line and tumors are of the same cancer type, using the other
  cell lines to derive ranks We use ranks to compare across cancer types, since the weights and
  score distributions are very different across data types and  cancer types
}
