% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_mean_similarity_to_gradient.R
\name{map_mean_similarity_to_gradient}
\alias{map_mean_similarity_to_gradient}
\title{Map mean similarity of cell lines to k nearest tumors to create a gradient from color 1 to color 2}
\usage{
map_mean_similarity_to_gradient(mean_similarity_cell_line_to_k_nearest_tumors,
  mean_similarity_tumor_to_k_nearest_tumors, col1 = "orange", col2 = "blue",
  numshades = 100)
}
\arguments{
\item{mean_similarity_cell_line_to_k_nearest_tumors}{a vector of mean similarities for each cell line to k nearest tumors}

\item{mean_similarity_tumor_to_k_nearest_tumors}{a vector of mean similarities for each tumor to k nearest tumors}

\item{col1}{a color (DEFAULT: orange)}

\item{col2}{a color (DEFAULT: blue)}

\item{numshades}{the number of shades on the color scale (DEFAULT: 100)}
}
\value{
a vector of colors of length equal to mean_similarity_cell_line_to_k_nearest_tumors
}
\description{
Map mean similarity of cell lines to k nearest tumors to create a gradient from color 1 to color 2
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{
tumorcomparer
}
