% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_dist.R
\name{pair_dist}
\alias{pair_dist}
\title{Calculate a weighted distance (between two vectors) based on Jaccard's coefficient}
\usage{
pair_dist(x, y, weights)
}
\arguments{
\item{x}{a vector of discrete values}

\item{y}{a vector of discrete values}

\item{weights}{a vector of non-negative alteration values}
}
\value{
a weighted distance based on Jaccard's coefficient. Ratio of the size 
  of intersection to the size of the union of two sets after discarding 0-0 
  matches from the two vectors
}
\description{
Calculate a weighted distance (between two vectors) based on Jaccard's coefficient
}
\examples{
n <- 100
x <- sample(c(0, 1), n, replace=TRUE)
y <- sample(c(0, 1), n, replace=TRUE)
weights <- rnorm(n)
pair_dist(x, y, weights)

}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
