% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balloon_plot.R
\name{plot_balloon_plot}
\alias{plot_balloon_plot}
\title{Make Balloon Plot for Results}
\usage{
plot_balloon_plot(dat, title)
}
\arguments{
\item{dat}{data generated from either make_}

\item{title}{a string to be the title for the plot}
}
\value{
a ggplot2 object
}
\description{
Make Balloon Plot for Results
}
\seealso{
\code{\link[=make_balloon_plot_data_from_comparison_result]{make_balloon_plot_data_from_comparison_result()}},
\code{\link[=make_balloon_plot_data_from_mtc]{make_balloon_plot_data_from_mtc()}}
}
\concept{tumorcomparer}
