% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq_alt.R
\name{plot_freq_alt}
\alias{plot_freq_alt}
\title{Plot mutation and copy number fractions}
\usage{
plot_freq_alt(num_cell_lines, num_tumors, composite_mut, composite_cna,
  tumor_color = "orange", cell_line_color = "blue", tumor_shape = 17,
  cell_line_shape = 20)
}
\arguments{
\item{num_cell_lines}{number of cell lines from FIX FUNCTION X}

\item{num_tumors}{number of cell lines from FIX FUNCTION Y}

\item{composite_mut}{a composite (both tumor and cell line information) matrix with only mutation information}

\item{composite_cna}{a composite (both tumor and cell line information) matrix with only copy number information}

\item{tumor_color}{a color for tumor points (DEFAULT: orange)}

\item{cell_line_color}{a color for tumor points (DEFAULT: blue)}

\item{tumor_shape}{an integer for an R plot PCH symbol}

\item{cell_line_shape}{an integer for an R plot PCH symbol}
}
\value{
Nothing is returned
}
\description{
Plot mutation and copy number fractions
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{
tumorcomparer
}

