% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_joyplot.R
\name{plot_joyplot}
\alias{plot_joyplot}
\title{Make Joy Plots for Distributions for Similarity Distributions}
\usage{
plot_joyplot(comparison_result)
}
\arguments{
\item{comparison_result}{the results of run_comparison() (See: run_comparison())}
}
\value{
a ggplot object
}
\description{
Make Joy Plots for Distributions for Similarity Distributions
}
\examples{
# Generated using: tumorcomparer::run_comparison() 
comparison_result <- readRDS(system.file("test_output", "ov_comparison_result.rds", 
  package="tumorcomparer"))
  
plot_joyplot(comparison_result = comparison_result)

}
