% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot the results from run_comparison as a two dimension multi-dimensional scaling (MDS) plot}
\usage{
plot_mds(
  comparison_list,
  categorization_list,
  trim_cell_line_names = FALSE,
  tumor_color = "blue",
  cell_line_color = "orange",
  use_gradient = TRUE,
  tumor_shape = 20,
  cell_line_shape = 17
)
}
\arguments{
\item{comparison_list}{the results of run_comparison() (See: run_comparison())}

\item{categorization_list}{the results of categorize_cell_lines() (See: categorize_cell_lines())}

\item{trim_cell_line_names}{a boolean whether to trim the the cell lines; 
this is optional and used for long cell line names in CCLE format (i.e. CELLLINE_TISSUE)}

\item{tumor_color}{a color for tumor points (DEFAULT: blue)}

\item{cell_line_color}{a color for tumor points (DEFAULT: orange)}

\item{use_gradient}{a boolean, cell lines will appear with a gradient of 
colors with those with the cell_line_color being least similar to the tumors 
and those being similar to tumors will have a color closer to the tumor_color.}

\item{tumor_shape}{an integer for an R plot PCH symbol (DEFAULT: 17)}

\item{cell_line_shape}{an integer for an R plot PCH symbol (DEFAULT: 20)}
}
\value{
Nothing is returned
}
\description{
Plot the results from run_comparison as a two dimension multi-dimensional scaling (MDS) plot
}
\seealso{
\code{\link{run_comparison}}, \code{\link{categorize_cell_lines}}
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
