% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comparison.R
\name{run_comparison}
\alias{run_comparison}
\title{Run a comparison between between two cohorts (e.g. cell lines and tumors)}
\usage{
run_comparison(
  available_data_types = c("mut", "cna", "exp"),
  mut_data_type_weight = 1/3,
  cna_data_type_weight = 1/3,
  exp_data_type_weight = 1/3,
  cna_default_weight = 0.01,
  mut_default_weight = 0.01,
  exp_default_weight = 0.01,
  tumor_mut_file,
  tumor_cna_file,
  tumor_exp_file,
  cell_line_mut_file,
  cell_line_cna_file,
  cell_line_exp_file,
  known_cancer_gene_weights_mut_file,
  known_cancer_gene_weights_cna_file,
  known_cancer_gene_weights_exp_file,
  cancer_specific_gene_weights_mut_file,
  cancer_specific_gene_weights_cna_file,
  cancer_specific_gene_weights_exp_file
)
}
\arguments{
\item{available_data_types}{a vector of data types to be analyzed}

\item{mut_data_type_weight}{a numeric weight for the data type (NOTE: data type weights must sum to 1)}

\item{cna_data_type_weight}{a numeric weight for the data type (NOTE: data type weights must sum to 1)}

\item{exp_data_type_weight}{a numeric weight for the data type (NOTE: data type weights must sum to 1)}

\item{cna_default_weight}{default (background) weight for copy number alterations (CNA) (DEFAULT: 0.01).
Default weights are assigned to genes not known to be important in the specific cancer type or cancer in general}

\item{mut_default_weight}{default (background) weight for mutation alterations (MUT) (DEFAULT: 0.01); CNA_default_weight}

\item{exp_default_weight}{default (background) weight for mRNA gene expression values (EXP) (DEFAULT: 0.01)}

\item{tumor_mut_file}{a file with binary mutation data for tumors}

\item{tumor_cna_file}{a file with GISTIC data for tumors; this can be 5-values (-2, -1, 0, 1, 2)}

\item{tumor_exp_file}{a file with gene expression data for tumors}

\item{cell_line_mut_file}{a file with binary mutation data for cell lines}

\item{cell_line_cna_file}{a file with GISTIC data for cell lines; this can be 5-values (-2, -1, 0, 1, 2)}

\item{cell_line_exp_file}{a file with gene expression data for cell lines}

\item{known_cancer_gene_weights_mut_file}{a file with weights for genes known
to be recurrently altered/mutated in cancer (e.g. recurrently mutated genes in TCGA pan-cancer analyses).
A two-column tab-delimited file - the first column has the gene names and the second column specifies the weights.}

\item{known_cancer_gene_weights_cna_file}{for copy number; see known_cancer_gene_weights_mut_file}

\item{known_cancer_gene_weights_exp_file}{for expression; see known_cancer_gene_weights_mut_file}

\item{cancer_specific_gene_weights_mut_file}{a file with weights for cancer-specific
set of recurrently mutated genes. A tab-delimited file - the first column has the gene names,
and the second column specifies the weights.}

\item{cancer_specific_gene_weights_cna_file}{for copy number; see cancer_specific_gene_weights_mut_file}

\item{cancer_specific_gene_weights_exp_file}{for expression; see cancer_specific_gene_weights_mut_file}
}
\value{
a list with multiple items. NOTE: The values of the dist and isomdsfit will depend on
parameter whether the values for a data type are discrete or continuous
\itemize{
\item dist_mat: a matrix of pairwise distances
\item isomdsfit: a two-column (2-dimension) fitting of the distances reduced to two dimensions via MDS - multidimensional scaling, using the isoMDS function
\item cor_unweighted: a matrix of unweighted pairwise correlations
\item composite_mat: the composite matrix (see Details)
\item cell_line_ids: a vector of cell line IDs/names with all data types
\item tumor_ids: a vector of tumor IDs with all data types
}
}
\description{
Run a comparison between between two cohorts (e.g. cell lines and tumors)
}
\examples{
tumor_mut_file <- system.file("extdata", "READ_data_for_running_TC", "tumor_mut.txt", 
  package="tumorcomparer")
tumor_cna_file <- system.file("extdata", "READ_data_for_running_TC", "tumor_cna.txt", 
  package="tumorcomparer")
tumor_exp_file <- system.file("extdata", "READ_data_for_running_TC", "tumor_exp.txt", 
  package="tumorcomparer")

cell_line_mut_file <- system.file("extdata", "READ_data_for_running_TC", "cell_line_mut.txt", 
  package="tumorcomparer")
cell_line_cna_file <- system.file("extdata", "READ_data_for_running_TC", "cell_line_cna.txt", 
  package="tumorcomparer")
cell_line_exp_file <- system.file("extdata", "READ_data_for_running_TC", "cell_line_exp.txt", 
  package="tumorcomparer")

known_cancer_gene_weights_mut_file <- system.file("extdata", "READ_data_for_running_TC", 
  "default_weights_for_known_cancer_genes_mut.txt", package="tumorcomparer")
known_cancer_gene_weights_cna_file <- system.file("extdata", "READ_data_for_running_TC", 
  "default_weights_for_known_cancer_genes_cna.txt", package="tumorcomparer")
known_cancer_gene_weights_exp_file <- system.file("extdata", "READ_data_for_running_TC", 
  "default_weights_for_known_cancer_genes_exp.txt", package="tumorcomparer")

cancer_specific_gene_weights_mut_file <- system.file("extdata", "READ_data_for_running_TC", 
  "Genes_and_weights_mut.txt", package="tumorcomparer")
cancer_specific_gene_weights_cna_file <- system.file("extdata", "READ_data_for_running_TC", 
  "Genes_and_weights_cna.txt", package="tumorcomparer")
cancer_specific_gene_weights_exp_file <- system.file("extdata", "READ_data_for_running_TC", 
  "Genes_and_weights_exp.txt", package="tumorcomparer")

comparison_result <- run_comparison(
available_data_types=c("mut", "cna", "exp"), 
mut_data_type_weight = 1/3,
cna_data_type_weight = 1/3,
exp_data_type_weight = 1/3,
cna_default_weight=0.01, 
mut_default_weight=0.01,
exp_default_weight=0.01,
tumor_mut_file=tumor_mut_file, 
tumor_cna_file=tumor_cna_file, 
tumor_exp_file=tumor_exp_file, 
cell_line_mut_file=cell_line_mut_file, 
cell_line_cna_file=cell_line_cna_file, 
cell_line_exp_file=cell_line_exp_file, 
known_cancer_gene_weights_mut_file=known_cancer_gene_weights_mut_file, 
known_cancer_gene_weights_cna_file=known_cancer_gene_weights_cna_file, 
known_cancer_gene_weights_exp_file=known_cancer_gene_weights_exp_file, 
cancer_specific_gene_weights_mut_file=cancer_specific_gene_weights_mut_file, 
cancer_specific_gene_weights_cna_file=cancer_specific_gene_weights_cna_file, 
cancer_specific_gene_weights_exp_file=cancer_specific_gene_weights_exp_file)

}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
