% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comparison_concatenated.R
\name{run_comparison_concatenated}
\alias{run_comparison_concatenated}
\title{Run a comparison between between two cohorts (e.g. cell lines and tumors)}
\usage{
run_comparison_concatenated(
  CNA_default_weight = 0.01,
  MUT_default_weight = 0.01,
  CNA_known_cancer_gene_weight = 0.1,
  MUT_known_cancer_gene_weight = 0.1,
  tumor_mut_file = "tumor_MUT.txt",
  tumor_cna_file = "tumor_CNA.txt",
  cell_line_mut_file = "cell_line_MUT.txt",
  cell_line_cna_file = "cell_line_CNA.txt",
  pancancer_gene_weights_file = "default_weights_for_known_cancer_genes.txt",
  cancer_specific_gene_weights_file = "Genes_and_weights.txt",
  output_composite_alteration_matrix_file = "composite_alteration_matrix.txt",
  distance_similarity_measure = c("weighted_correlation", "generalized_jaccard")
)
}
\arguments{
\item{CNA_default_weight}{default weight for copy number alterations (CNA) (DEFAULT: 0.01)}

\item{MUT_default_weight}{default weight for mutation alterations (MUT) (DEFAULT: 0.01)}

\item{CNA_known_cancer_gene_weight}{a default weight (DEFAULT: 0.1) - genes in 
the TCGA pan-cancer CNA list (i.e. GISTIC peaks)}

\item{MUT_known_cancer_gene_weight}{a default weight (DEFAULT: 0.1) - genes in
the TCGA pan-cancer mutation list (i.e. MUTSIG SMGs)}

\item{tumor_mut_file}{a file with binary mutation data for tumors, (e.g. over the 
1651 genes profiled by CCLE)}

\item{tumor_cna_file}{a file with 5-valued GISTIC data for tumors, (e.g. over 
1529 genes (subset of 1651 genes above))}

\item{cell_line_mut_file}{a file with binary mutation data for cell lines, 
(e.g. over the 1651 genes profiled by CCLE)}

\item{cell_line_cna_file}{a file with 5-valued GISTIC data for cell lines, 
(e.g. over 1529 genes (subset of 1651 genes above))}

\item{pancancer_gene_weights_file}{- a file with weights for the TCGA pan-cancer
set of recurrent mutations and CNAs (copy number alterations). A two-column
tab-delimited file - the first column has the list of alterations, with "_MUT"
and "_CNA" specifying mutations and CNAs(copy number alterations) respectively,
e.g. "TP53_MUT" and "MYC_CNA", and the second column specifying the weights.}

\item{cancer_specific_gene_weights_file}{a file with weights for cancer-specific
set of recurrent mutations and CNAs (copy number alterations). A two-column
tab-delimited file - the first column has the list of alterations, with "_MUT"
and "_CNA" specifying mutations and CNAs(copy number alterations) respectively,
e.g. "TP53_MUT" and "MYC_CNA", and the second column specifying the weights.}

\item{output_composite_alteration_matrix_file}{a string with filename for 
output of the the composite alteration matrix (see Details) 
(DEFAULT: "composite_alteration_matrix.txt")}

\item{distance_similarity_measure}{(See Details) 
(OPTIONS: "weighted_correlation", "generalized_jaccard")}
}
\value{
a list with multiple items. NOTE: The values of the dist and isomdsfit will
 depend on parameter: distance_similarity_measure.
\itemize{
  \item{"dist"}{a matrix of pairwise distances}
  \item{"isomdsfit"}{a two-column, 2-dimension, fitting of the distances reduced to two dimensions via MDS - multidimensional scaling, using the isoMDS function}
  \item{"cor_unweighted"}{a matrix of unweighted pairwise correlations}
  \item{"composite_mat"}{the composite matrix, see Details}
  \item{"cell_lines_with_both_MUT_and_CNA"}{a vector of cell line IDs/names with both mutation
   (MUT) and copy number alteration (CNA) information}
  \item{"tumors_with_both_MUT_and_CNA"}{a vector of tumor IDs with both mutation 
  (MUT) and copy number alteration (CNA) information}
}
}
\description{
Run a comparison between between two cohorts (e.g. cell lines and tumors)
}
\details{
The composite matrix is a single matrix where the columns are samples 
  (i.e. tumors AND cell line IDs) and the rows are an rbind() of mutations 
  (with 1 or 0 outputs for each sample) and copy number alterations from 
  GISTIC (with values -2, -1, 0, 1, 2). Available similarity/distance measures include: 
  \itemize{
    \item{"weighted_correlation"}{Weighted correlation, based on weighted means and standard deviations}
    \item{"generalized_jaccard"}{A weighted distance based on the Jaccard coefficient}
 }
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
