% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comparison_config_list.R
\name{run_comparison_config_list}
\alias{run_comparison_config_list}
\title{FIXME}
\usage{
run_comparison_config_list(
  config_list,
  gene_list = NULL,
  remove_errored_dataset_comparisons = FALSE
)
}
\arguments{
\item{config_list}{FIXME}

\item{gene_list}{a vector of HGNC gene symbols to run comparison only for the specified genes (Default: NULL)}

\item{remove_errored_dataset_comparisons}{FIXME (Default: FALSE)}
}
\value{
FIXME
}
\description{
FIXME
}
