% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_geneset_comparison.R
\name{tc_geneset_comparison}
\alias{tc_geneset_comparison}
\title{Gene Set Comparison Function for TCGA Dataset}
\usage{
tc_geneset_comparison(
  gene_list,
  cancer_type = NULL,
  tcga_ids = NULL,
  cclp_ids = NULL,
  tc_dataset = NULL,
  tc_dataset_dir = NULL,
  remove_errored_dataset_comparisons = FALSE,
  remove_tmp_files = TRUE,
  run_mds = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{gene_list}{a vector of HGNC gene symbols which will be used to perform the comparison}

\item{cancer_type}{TCGA cancer name (GBM, LGG, etc.)}

\item{tcga_ids}{a vector of TCGA IDs for analysis; cannot be combined with cancer_type}

\item{cclp_ids}{a vector of CCLP IDs for analysis; cannot be combined with cancer_type}

\item{tc_dataset}{path to RData for TumorComparer dataset from publication; see details}

\item{tc_dataset_dir}{path to directory with TumorComparer dataset as text files (recommended instead of tc_dataset); see details}

\item{remove_errored_dataset_comparisons}{will skip the data types which can't be compared for technical reasons(no enaugh genes to compare, or data contain only 0 values) when set to TRUE (Default: FALSE)}

\item{remove_tmp_files}{remove temporary files (Default: TRUE)}

\item{run_mds}{a boolean, whether to run multidimensional scaling (MDS) on dataset (Default: TRUE)}

\item{verbose}{show debugging information}
}
\value{
returns comparison list for specified genes. The output list is described in run_comparison_config_list function documentation
}
\description{
Gene Set Comparison Function for TCGA Dataset
}
\details{
This function requires downloading: https://zenodo.org/record/4627644/files/tc_data_pancancer_march2021.tar.gz?download=1

The above data is the dataset used in the 2021 TumorComparer publication.
Data can then be loaded from the extracted text files if 'tc_dataset_dir'
parameter is provided or data can be loaded from an RData workspace variable
using the 'tc_dataset' parameter if it has been pre-generated. tc_dataset_dir'
is recommended.
}
\note{
FIXME ADD EXAMPLE
}
