% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_geneset_comparison.R
\name{tcga_geneset_comparison}
\alias{tcga_geneset_comparison}
\title{geneset comparison function for TCGA dataset}
\usage{
tcga_geneset_comparison(
  tcga_dataset,
  cancer_type,
  gene_list,
  remove_errored_dataset_comparisons = FALSE
)
}
\arguments{
\item{tcga_dataset}{path to the tcga dataset (tcga_comparison_data.RData), see details}

\item{cancer_type}{TCGA cancer name (GBM, LGG, etc.)}

\item{gene_list}{a vector of HGNC gene symbols which will be used to perform the comparison}

\item{remove_errored_dataset_comparisons}{will skip the data types which can't be compared for technical reasons(no enaugh genes to compare, or data contain only 0 values) when set to TRUE (Default: FALSE)}
}
\value{
returns comparison list for specified genes. The output list is described in run_comparison_config_list function documentation
}
\description{
geneset comparison function for TCGA dataset
}
\details{
For running tcga_geneset_comparison you will need to download tcga_comparison_data.RData data file from ... and specify the path of the RData file in tcga_dataset argument
}
