% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_pancan_pathway_genes}
\alias{tcga_pancan_pathway_genes}
\title{Genes for TCGA Pan-Cancer Pathways}
\format{A list of vectors}
\source{
\doi{10.1016/j.cell.2018.03.035}
}
\usage{
tcga_pancan_pathway_genes
}
\description{
A dataset containing a list of genes for the TCGA Pan-Cancer Pathways paper (Sanchez-Vega et al 2018)
}
\keyword{datasets}
