% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic_toc.R
\name{tic}
\alias{tic}
\title{Starts a stopwatch timer to measure performance}
\usage{
tic(gcFirst = TRUE, type = c("elapsed", "user.self", "sys.self"))
}
\arguments{
\item{gcFirst}{a boolean whether to run garbage collection before starting watch}

\item{type}{the time of time to keep track of}
}
\description{
Starts a stopwatch timer to measure performance
}
\concept{tumorcomparer}
