% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weighted_corr.R
\name{calc_weighted_corr}
\alias{calc_weighted_corr}
\title{Calculate weighted correlation}
\usage{
calc_weighted_corr(a, b, w = rep(1, nrow(a))/nrow(a))
}
\arguments{
\item{a}{numeric matrix}

\item{b}{numeric matrix}

\item{w}{alteration weights}
}
\value{
a weighted correlation (similarity) matrix
}
\description{
Calculate weighted correlation
}
\examples{
t1 <- as.matrix(c(1,1,1,1,1,1,2,2,2,3,3,3,4,4))
t2 <- as.matrix(rev(t1))
weight <- c(.5,.5,.5,.5,.5,1,1,1,1,2,2,2,2,2)
calc_weighted_corr(t1, t2, weight)

}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
