% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_cell_lines.R
\name{categorize_cell_lines}
\alias{categorize_cell_lines}
\title{Categorize cell lines by the level of similarity [FIX]}
\usage{
categorize_cell_lines(fraction_of_tumors_for_comparison = 0.1, dist_mat,
  composite_mat, cell_lines_with_both_MUT_and_CNA, tumors_with_both_MUT_and_CNA)
}
\arguments{
\item{fraction_of_tumors_for_comparison}{fraction of tumors used in a k-nearest neighbor comparison? (DEFAULT: 0.1) [FIX]}

\item{dist_mat}{a mat [FIX]}

\item{composite_mat}{a mat [FIX]}

\item{cell_lines_with_both_MUT_and_CNA}{a vector [FIX]}

\item{tumors_with_both_MUT_and_CNA}{a vector [FIX]}
}
\description{
Categorize cell lines by the level of similarity [FIX]
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{
tumorcomparer
}

