% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_cell_lines.R
\name{categorize_cell_lines}
\alias{categorize_cell_lines}
\title{Categorize cell lines by the level of similarity to k-nearest tumors (DEFAULT: 10 percent of tumors)}
\usage{
categorize_cell_lines(num_tumors_for_comparison = 10, dist_mat, cell_line_ids,
  tumor_ids, trim_cell_line_names = FALSE)
}
\arguments{
\item{num_tumors_for_comparison}{number of tumors used in a k-nearest
neighbor comparison (DEFAULT: 10)}

\item{dist_mat}{a matrix of pairwise weighted distances between all cell lines and tumors}

\item{cell_line_ids}{IDs/names of cell lines}

\item{tumor_ids}{IDs of tumors}

\item{trim_cell_line_names}{a boolean whether to trim the the cell lines; 
this is optional and used for long cell line names in CCLE format (i.e. CELLLINE_TISSUE); (DEFAULT: FALSE)}

\item{composite_mat}{a composite matrix of alterations in tumors and cell lines 
(see "get_tumor_comparison.R" for detail)}
}
\value{
a list with the following items: 
\itemize{
  \item{"mean_similarity_cell_line_to_k_nearest_tumors"}{the mean similarity of each cell line to the k-nearest tumors calculated using k=fraction_of_tumors_for_comparison*number of tumors}
  \item{"mean_similarity_tumor_to_k_nearest_tumors"}{the mean similarity of each tumor sample to the k-nearest tumors calculated using k=fraction_of_tumors_for_comparison*number of tumors}
  \item{"categorization"}{a 2-column data.frame with the cell line categorizations: Sample_ID and Category; Category values can be: "Great", "Good", "Moderately Good", "Poor", "Outliers"}
}
}
\description{
Categorize cell lines by the level of similarity to k-nearest tumors (DEFAULT: 10 percent of tumors)
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{
tumorcomparer
}
