% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_composite_mat_and_gene_weights.R
\name{generate_composite_mat_and_gene_weights}
\alias{generate_composite_mat_and_gene_weights}
\title{Run a comparison between between two cohorts (e.g. cell lines and tumors)}
\usage{
generate_composite_mat_and_gene_weights(
  default_weight,
  tumor_file,
  cell_line_file,
  known_cancer_gene_weights_file = NULL,
  cancer_specific_gene_weights_file = NULL
)
}
\arguments{
\item{default_weight}{see run_comparison}

\item{tumor_file}{see run_comparison}

\item{cell_line_file}{see run_comparison}

\item{known_cancer_gene_weights_file}{see run_comparison}

\item{cancer_specific_gene_weights_file}{see run_comparison}
}
\value{
a list with multiple items. NOTE: The values of the dist and isomdsfit will
depend on whether the input data is continuous or discrete.
\itemize{
\item{"dist_mat"}{a matrix of pairwise distances}
\item{"isomdsfit"}{a two-column (2-dimension) fitting of the distances reduced to
two dimensions via MDS - multidimensional scaling, using the isoMDS function}
\item{"cor_unweighted"}{a matrix of unweighted pairwise correlations}
\item{"composite_mat"}{the composite matrix (see Details)}
\item{"cell_lines_ids"}{a vector of cell line IDs/names}
\item{"tumors_ids"}{a vector of tumor IDs}
}
}
\description{
Run a comparison between between two cohorts (e.g. cell lines and tumors)
}
\details{
The composite matrix is a single matrix where the columns are samples
(i.e. tumors AND cell line IDs) and the rows are an rbind() of mutations
(with 1 or 0 outputs for each sample), copy number alterations from
GISTIC (with values -2, -1, 0, 1, 2), or gene expression values. Available similarity/distance measures include:
\itemize{
\item{"weighted_correlation"}{Weighted correlation, based on weighted means and standard deviations}
\item{"generalized_jaccard"}{A weighted distance based on the Jaccard coefficient}
}
}
\examples{
mut_default_weight <- 0.01
tumor_mut_file <- system.file("extdata", "READ_data_for_running_TC", "tumor_mut.txt", 
  package="tumorcomparer")
cell_line_mut_file <- system.file("extdata", "READ_data_for_running_TC", "cell_line_mut.txt", 
  package="tumorcomparer")
known_cancer_gene_weights_mut_file <- system.file("extdata", "READ_data_for_running_TC", 
  "default_weights_for_known_cancer_genes_mut.txt", package="tumorcomparer")
cancer_specific_gene_weights_mut_file <- system.file("extdata", "READ_data_for_running_TC", 
  "Genes_and_weights_mut.txt", package="tumorcomparer")

mut <- generate_composite_mat_and_gene_weights(
  default_weight=mut_default_weight,
  tumor_file=tumor_mut_file,
  cell_line_file=cell_line_mut_file,
  known_cancer_gene_weights_file=known_cancer_gene_weights_mut_file,
  cancer_specific_gene_weights_file=cancer_specific_gene_weights_mut_file)

}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
