% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphNELtoJSON.R
\name{graphNELtoJSON}
\alias{graphNELtoJSON}
\title{Convert R graphNEL object to cytoscape.js JSON.}
\usage{
graphNELtoJSON(g)
}
\arguments{
\item{g}{a graphNEL}
}
\description{
Convert R graphNEL object to cytoscape.js JSON.
}
\concept{tumorcomparer}
