% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_dist.R
\name{pair_dist}
\alias{pair_dist}
\title{Calculate a weighted distance (between two vectors) based on Jaccard's coefficient}
\usage{
pair_dist(x, y, weights)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{weights}{alteration weights}
}
\value{
a weighted distance based on Jaccard's coefficient (ratio of the size of intersection to the size of union of two sets)
}
\description{
Calculate a weighted distance (between two vectors) based on Jaccard's coefficient
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
