% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balloon_plot.R
\name{plot_balloon_plot}
\alias{plot_balloon_plot}
\title{Make Balloon Plot for Results}
\usage{
plot_balloon_plot(dat, title)
}
\value{
a ggplot2 object
}
\description{
Make Balloon Plot for Results
}
\concept{tumorcomparer}
