% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq_alt.R
\name{plot_freq_alt}
\alias{plot_freq_alt}
\title{Plot mutation and copy number fractions}
\usage{
plot_freq_alt(comparison_list, tumor_color = "blue",
  cell_line_color = "orange", tumor_shape = 17, cell_line_shape = 20,
  trim_cell_line_names = TRUE, max_xlim = 1)
}
\arguments{
\item{comparison_list}{the results of run_comparison() (See: run_comparison())}

\item{tumor_color}{a color for tumor points (DEFAULT: blue)}

\item{cell_line_color}{a color for tumor points (DEFAULT: orange)}

\item{tumor_shape}{an integer for an R plot PCH symbol}

\item{cell_line_shape}{an integer for an R plot PCH symbol}

\item{trim_cell_line_names}{a boolean whether to trim the the cell lines; 
this is optional and used for long cell line names in CCLE format (i.e. CELLLINE_TISSUE)}

\item{max_xlim}{a value between 0 and 1 the maximum value to for x-axis}
}
\value{
Nothing is returned
}
\description{
Plot mutation and copy number fractions
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{
tumorcomparer
}

