% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{Run a comparison between}
\usage{
plot_mds(cell_lines_with_both_MUT_and_CNA, tumors_with_both_MUT_and_CNA,
  dist_mat, cell_line_ids = NULL, tumor_color = "orange",
  cell_line_color = "blue", tumor_shape = 17, cell_line_shape = 20)
}
\arguments{
\item{dist_mat}{a matrix of distances}

\item{cell_line_ids}{a vector of strings with cell line names. This will be taken from
cell_lines_with_both_MUT_and_CNA and assumed to be in the CCLE format (i.e. CELLLINE_TISSUE)}

\item{tumor_color}{a color for tumor points (DEFAULT: orange)}

\item{cell_line_color}{a color for tumor points (DEFAULT: blue)}

\item{tumor_shape}{an integer for an R plot PCH symbol (DEFAULT: 17)}

\item{cell_line_shape}{an integer for an R plot PCH symbol (DEFAULT: 20)}

\item{composite_mut}{a composite (both tumor and cell line information) matrix 
with only mutation information}

\item{composite_cna}{a composite (both tumor and cell line information) matrix 
with only copy number information}
}
\value{
Nothing is returned
}
\description{
Run a comparison between
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{
tumorcomparer
}

