% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comparison.R
\name{run_comparison}
\alias{run_comparison}
\title{Run a comparison between between two cohorts (e.g. cell lines and tumors)}
\usage{
run_comparison(
  available_data_types = c("mut", "cna", "exp"),
  mut_data_type_weight = 1/3,
  cna_data_type_weight = 1/3,
  exp_data_type_weight = 1/3,
  cna_default_weight = 0.01,
  mut_default_weight = 0.01,
  exp_default_weight = 0.01,
  cna_known_cancer_gene_weight = 0.1,
  mut_known_cancer_gene_weight = 0.1,
  exp_known_cancer_gene_weight = 0.1,
  tumor_mut_file = "tumor_mut.txt",
  tumor_cna_file = "tumor_cna.txt",
  tumor_exp_file = "tumor_exp.txt",
  cell_line_mut_file = "cell_line_mut.txt",
  cell_line_cna_file = "cell_line_cna.txt",
  cell_line_exp_file = "cell_line_exp.txt",
  known_cancer_gene_weights_mut_file = "default_weights_for_known_cancer_genes_mut.txt",
  known_cancer_gene_weights_cna_file = "default_weights_for_known_cancer_genes_cna.txt",
  known_cancer_gene_weights_exp_file = "default_weights_for_known_cancer_genes_exp.txt",
  cancer_specific_gene_weights_mut_file = "Genes_and_weights_mut.txt",
  cancer_specific_gene_weights_cna_file = "Genes_and_weights_cna.txt",
  cancer_specific_gene_weights_exp_file = "Genes_and_weights_exp.txt",
  distance_similarity_measures = c("generalized_jaccard", "weighted_correlation",
    "weighted_correlation")
)
}
\arguments{
\item{available_data_types}{a vector of data types to be analyzed; the order of these much match distance_similarity_measures}

\item{mut_data_type_weight}{TODO}

\item{cna_data_type_weight}{TODO}

\item{exp_data_type_weight}{TODO}

\item{cna_default_weight}{default (background) weight for copy number alterations (CNA) (DEFAULT: 0.01).
Default weights are assigned to genes not known to be important in the specific cancer type or cancer in general}

\item{mut_default_weight}{default (background) weight for mutation alterations (MUT) (DEFAULT: 0.01); CNA_default_weight}

\item{exp_default_weight}{default (background) weight for mRNA gene expression values (EXP) (DEFAULT: 0.01)}

\item{cna_known_cancer_gene_weight}{a default weight for genes important in cancer (DEFAULT: 0.1);
an example would genes in the TCGA pan-cancer CNA list (i.e. GISTIC peaks). Known cancer gene weights are
default weight for genes important in cancer (an example would genes derived from TCGA pan-cancer analyses or literature)}

\item{mut_known_cancer_gene_weight}{a default weight for genes important in cancer (DEFAULT: 0.1);
an example would genes in the TCGA pan-cancer mutation list (i.e. MUTSIG SMGs)}

\item{exp_known_cancer_gene_weight}{a default weight for genes important in cancer(DEFAULT: 0.1);
an example would genes in the COSMIC Cancer Gene Census}

\item{tumor_mut_file}{a file with binary mutation data for tumors}

\item{tumor_cna_file}{a file with 5-valued GISTIC data for tumors}

\item{tumor_exp_file}{a file with gene expression data for tumors}

\item{cell_line_mut_file}{a file with binary mutation data for cell lines}

\item{cell_line_cna_file}{a file with 5-valued GISTIC data for cell lines}

\item{cell_line_exp_file}{a file with gene expression data for cell lines}

\item{known_cancer_gene_weights_mut_file}{a file with weights for genes known
to be recurrently altered/mutated in cancer (e.g. recurrently mutated genes in TCGA pan-cancer analyses).
A two-column tab-delimited file - the first column has the gene names and the second column specifies the weights.}

\item{known_cancer_gene_weights_cna_file}{see known_cancer_gene_weights_mut_file}

\item{known_cancer_gene_weights_exp_file}{see known_cancer_gene_weights_mut_file}

\item{cancer_specific_gene_weights_mut_file}{a file with weights for cancer-specific
set of recurrently mutated genes. A tab-delimited file - the first column has the gene names,
and the second column specifies the weights.}

\item{cancer_specific_gene_weights_cna_file}{see cancer_specific_gene_weights_mut_file}

\item{cancer_specific_gene_weights_exp_file}{see cancer_specific_gene_weights_mut_file}

\item{distance_similarity_measures}{a named vector of distance/similarity measures: see details.
OPTIONS: weighted_correlation and generalized_jaccard - must be in the order mut, cna, exp. Currently,
generalized_jaccard is used for mut and cna data, and weighted_correlation for exp data}
}
\value{
a list with multiple items. NOTE: The values of the dist and isomdsfit will depend on parameter, distance_similarity_measure.
\itemize{
\item dist_mat: a matrix of pairwise distances
\item isomdsfit: a two-column (2-dimension) fitting of the distances reduced to two dimensions via MDS - multidimensional scaling, using the isoMDS function
\item cor_unweighted: a matrix of unweighted pairwise correlations
\item composite_mat: the composite matrix (see Details)
\item cell_line_ids: a vector of cell line IDs/names with all data types
\item tumor_ids: a vector of tumor IDs with all data types
}
}
\description{
Run a comparison between between two cohorts (e.g. cell lines and tumors)
}
\author{
Rileen Sinha (rileen@gmail.com), Augustin Luna (aluna@jimmy.harvard.edu)
}
\concept{tumorcomparer}
